requirements.txt
Django==3.2.15
django-environ==0.8.1
gunicorn==20.1.0
uvicorn
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django_cas_ng
